// frontend/src/app/learncoreservices/models/category-response.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface CategoryResponse {
  description?: string;
  imageUrl?: string;
  languageCode?: string;
  languageUuid?: string;
  name?: string;
  parentName?: string;
  parentUuid?: string;
  slug?: string;
  uuid?: string;
}